move week part of calendar into its own file
